refactor json config

+++ ext-markdown-math/src/main/java/me/chan/texas/ext/markdown/math/renderer/fonts/Symbol.java
@@ -2,12 +2,12 @@ package me.chan.texas.ext.markdown.math.renderer.fonts;
 
 public class Symbol {
 	public final String unicode;
-	public final int xMin;
-	public final int ascent;
-	public final int xMax;
-	public final int descent;
+	public final float xMin;
+	public final float ascent;
+	public final float xMax;
+	public final float descent;
 
-	public Symbol(String unicode, int xMin, int ascent, int xMax, int descent) {
+	public Symbol(String unicode, float xMin, float ascent, float xMax, float descent) {
 		this.unicode = unicode;
 		this.xMin = xMin;
 		this.ascent = ascent;
